test: split up #1331 tests
